packages/mcp-server-hosted/pyproject.toml
[project]
name = "finout-mcp-hosted"
version = "0.1.0"
description = "Hosted public MCP server with Frontegg OAuth — internal deployment only"
requires-python = ">=3.11"
dependencies = [
    "finout-mcp",
    "PyJWT[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "starlette>=0.37.0",
    "uvicorn>=0.30.0",
    "anyio>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
finout-mcp-hosted = "finout_mcp_hosted.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finout_mcp_hosted"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
